main/Streamlit-based Intelligent Learning Platform/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
easyocr==1.7.2
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gevent==24.10.2
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
ipython==8.12.3
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.11
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
openai==1.51.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
paddleocr==2.9.0
paddlepaddle==2.6.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.2
pure_eval==0.2.3
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.2.0
shap==0.46.0
shapely==2.0.6
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spark-ai-python==0.4.5
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
streamlink==6.11.0
streamlit==1.39.0
streamlit-chat==0.1.1
sympy==1.13.3
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.15.4
zope.event==5.0
zope.interface==7.1.0
